Refactor BIRFunction class
